# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "Simple CRUD API for tasks."
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "paths": {
    "/tasks": {
      "get": {
        "summary": "List all tasks",
        "responses": {
          "200": {
            "description": "Tasks returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Task" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NewTask" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "400": { "description": "Title is required" }
        }
      }
    },
    "/tasks/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Task ID",
          "schema": { "type": "integer" }
        }
      ],
      "get": {
        "summary": "Get a task by id",
        "responses": {
          "200": {
            "description": "Task returned",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "404": { "description": "Task not found" }
        }
      },
      "put": {
        "summary": "Update a task by id",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/TaskUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "400": { "description": "Request body cannot be empty" },
          "404": { "description": "Task not found" }
        }
      },
      "delete": {
        "summary": "Delete a task by id",
        "responses": {
          "200": {
            "description": "Task deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeleteResult" }
              }
            }
          },
          "404": { "description": "Task not found" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Task": {
        "type": "object",
        "required": ["id", "title", "done"],
        "properties": {
          "id": { "type": "integer", "example": 1 },
          "title": { "type": "string", "example": "Learn Express basics" },
          "done": { "type": "boolean", "example": false }
        }
      },
      "NewTask": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string", "example": "Write API docs" }
        }
      },
      "TaskUpdate": {
        "type": "object",
        "minProperties": 1,
        "properties": {
          "title": { "type": "string" },
          "done": { "type": "boolean" }
        }
      },
      "DeleteResult": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "Task deleted" }
        }
      }
    }
  }
}